Move `AuthorizationsController` to decidim-verifications
Add a new `Decidim::Verifications::Engine` that mounts all available
authorization workflows.
