+++ decidim-core/app/controllers/decidim/devise/sessions_controller.rb
@@ -7,9 +7,11 @@ module Decidim
       include Decidim::DeviseControllers
 
       def after_sign_in_path_for(user)
-        return first_login_authorizations_path if first_login_and_not_authorized?(user) &&
-                                                  !user.admin?
-        super
+        if first_login_and_not_authorized?(user) && !user.admin?
+          decidim_verifications.first_login_authorizations_path
+        else
+          super
+        end
       end
 
       def first_login_and_not_authorized?(user)
