+++ decidim-core/config/locales/en.yml
@@ -77,23 +77,6 @@ en:
         name: Foo authorization
       granted_at: Granted at %{timestamp}
       started_at: Started at %{timestamp}
-    authorizations:
-      create:
-        error: There was an error creating the authorization.
-        success: You've been successfully authorized.
-      first_login:
-        actions:
-          decidim/dummy_authorization_handler: Verify against the example authorization handler
-          dummy_authorization_workflow: Verify against the example authorization workflow
-          id_documents: Get verified by uploading your identity document
-          postal_letter: Get verified by receiving a verification code through postal mail
-        title: Verify your identity
-        verify_with_these_options: 'These are the available options to verify your identity:'
-      new:
-        authorize: Send
-        authorize_with: Verify with %{authorizer}
-      skip_verification: You can skip this for now and %{link}
-      start_exploring: start exploring
     core:
       actions:
         unauthorized: You are not authorized to perform this action
