+++ decidim-verifications/app/controllers/decidim/verifications/postal_letter/authorizations_controller.rb
@@ -22,7 +22,7 @@ module Decidim
           PerformAuthorizationStep.call(@authorization, @form) do
             on(:ok) do
               flash[:notice] = t("authorizations.create.success", scope: "decidim.verifications.postal_letter")
-              redirect_to decidim.authorizations_path
+              redirect_to decidim_verifications.authorizations_path
             end
 
             on(:invalid) do
@@ -46,7 +46,7 @@ module Decidim
           ConfirmUserAuthorization.call(@authorization, @form) do
             on(:ok) do
               flash[:notice] = t("authorizations.update.success", scope: "decidim.verifications.postal_letter")
-              redirect_to decidim.authorizations_path
+              redirect_to decidim_verifications.authorizations_path
             end
 
             on(:invalid) do
