+++ decidim-verifications/app/views/decidim/verifications/authorizations/index.html.erb
