+++ decidim-verifications/app/views/decidim/verifications/authorizations/new.html.erb
@@ -23,7 +23,7 @@
               <% end %>
             <% end %>
             <p class="text-center skip">
-              <%= t("decidim.authorizations.skip_verification", link: link_to(t("decidim.authorizations.start_exploring"), cta_button_path).html_safe).html_safe %>.
+              <%= t("decidim.verifications.authorizations.skip_verification", link: link_to(t("decidim.verifications.authorizations.start_exploring"), cta_button_path).html_safe).html_safe %>.
             </p>
           </div>
         </div>
