+++ decidim-verifications/spec/controllers/decidim/authorizations_controller_spec.rb
@@ -2,9 +2,9 @@
 
 require "spec_helper"
 
-module Decidim
+module Decidim::Verifications
   describe AuthorizationsController, type: :controller do
-    routes { Decidim::Core::Engine.routes }
+    routes { Decidim::Verifications::Engine.routes }
 
     let(:user) { create(:user, :confirmed) }
